+++ lib/routes/app_route.dart
@@ -1,8 +1,7 @@
 import 'package:get/get.dart';
 import 'package:news_api/ui/country/view/country_view.dart';
 import 'package:news_api/ui/home/view/home_view.dart';
-
-import '../ui/splash/view/splash_view.dart';
+import 'package:news_api/ui/splash/view/splash_view.dart';
 
 const kDefaultTransition = Transition.rightToLeft;
 const kDefaultTransitionDuration = Duration(milliseconds: 500);