+++ lib/ui/country/model/country_model.dart
@@ -16,8 +16,8 @@ class Country {
 
 // List of countries
 final countries = [
-  Country(name: "India", code: "in", flagPath: "assets/svg/ic_india.svg"),
   Country(name: "United States", code: "us", flagPath: "assets/svg/ic_usa.svg"),
+  Country(name: "India", code: "in", flagPath: "assets/svg/ic_india.svg"),
   Country(name: "Canada", code: "ca", flagPath: "assets/svg/ic_canada.svg"),
   Country(name: "United Kingdom", code: "uk", flagPath: "assets/svg/ic_uk.svg"),
   Country(name: "Brazil", code: "br", flagPath: "assets/svg/ic_brazil.svg"),