+++ lib/ui/splash/controller/splash_controller.dart
@@ -20,7 +20,6 @@ class SplashController extends GetxController {
       // Get.to(() => HomeView()); class based approach
 
       Get.toNamed(HomeView.name); // name based approach
-      
     });
   }